requirements_ffmpeg.txt
ffmpeg_python>=0.2.0
Pillow>=11.0.0
rich>=13.9.4
